layout and tablet rendering refactoring - removed postprocessing and use finalizer instead - the layout info can be computed during present for most elements - when requiring layout or render buffer, call finalize - otherwise finalize is at least called once per present - added concept of section and section root - section root is responsible for finalizing its sub elements - section root also can handle rendering of sub elements collectively - finalizing is a process to set attributes based on the defined ones - mostly it's used to compute layout - added scalar which contains a number with additional semantic flags - such value can convey whether it's undefined, relative etc. - tablet rendering now take variable number of glyphs - each glyph carries data like position and dimensions

+++ source/game/game.cpp
@@ -48,8 +48,8 @@ void Game::present(const Context& ctx)
 	vp.pose.pos.y = ((float)tablet_size.y - vp_height) / 2;
 
 	viewport(_ctx);
-	_attr(attrs::width, static_cast<double>(window->width()));
-	_attr(attrs::height, static_cast<double>(window->height()));
+	_attr(attrs::width, window->width());
+	_attr(attrs::height, window->height());
 	_attr(attrs::viewpoint, vp);
 	_attr(attrs::background_color, Color{});
